Split Driver & LineSensor
