+++ src/linesensor.h
@@ -1,54 +1,22 @@
 #pragma once
 
 #include <vector>
-#include <atomic>
-
 #include <driver/spi_master.h>
+
+#include "driver.h"
 
 namespace mcp3008 {
 
 class LineSensorCalibrator;
 
 /**
- * \brief The LineSensor device - one MCP3008 chip.
+ * \brief The LineSensor device, based on MCP3008 chip.
  *
  * This class is not thread-safe, you have to make sure the methods are called
  * from one thread at a time only.
- * The install() method has to be called before you can use any other methods.
  */
-class LineSensor {
+class LineSensor : public Driver {
 public:
-    static constexpr int CHANNELS = 8; //!< Amount of channels on the chip
-    static constexpr uint16_t MAX_VAL = 1023; //!< Maximum value returned by from the chip (10bits).
-
-    /**
-     * \brief The LineSensor SPI configuration.
-     */
-    struct Config {
-        Config(gpio_num_t pin_cs = GPIO_NUM_25, gpio_num_t pin_mosi = GPIO_NUM_33,
-            gpio_num_t pin_miso = GPIO_NUM_32, gpio_num_t pin_sck = GPIO_NUM_26,
-            uint8_t channels_mask = 0xFF, int freq = 1350000, spi_host_device_t spi_dev = HSPI_HOST) {
-            this->freq = freq;
-            this->spi_dev = spi_dev;
-            this->channels_mask = channels_mask;
-
-            this->pin_cs = pin_cs;
-            this->pin_mosi = pin_mosi;
-            this->pin_miso = pin_miso;
-            this->pin_sck = pin_sck;
-        }
-
-        int freq; //!< SPI communication frequency
-        spi_host_device_t spi_dev; //!< Which ESP32 SPI device to use.
-        uint8_t channels_mask; //!< Which channels to use, bit mask:
-                               //!< (1 << 0) | (1 << 2) == channels 0 and 2 only.
-
-        gpio_num_t pin_cs;
-        gpio_num_t pin_mosi;
-        gpio_num_t pin_miso;
-        gpio_num_t pin_sck;
-    };
-
     /**
      * \brief The LineSensor's calibration data
      *
@@ -57,38 +25,18 @@ public:
      */
     struct CalibrationData {
         CalibrationData() {
-            for(int i = 0; i < CHANNELS; ++i) {
+            for(int i = 0; i < Driver::CHANNELS; ++i) {
                 min[i] = 0;
-                range[i] = MAX_VAL;
+                range[i] = Driver::MAX_VAL;
             }
         }
 
-        uint16_t min[CHANNELS];
-        uint16_t range[CHANNELS];
+        uint16_t min[Driver::CHANNELS];
+        uint16_t range[Driver::CHANNELS];
     };
 
     LineSensor();
-    virtual ~LineSensor(); //!< The uninstall() method is called from the destructor.
-
-    /**
-     * \brief Initialize the SPI bus. Must be called before any other methods,
-     *        otherwise they will return ESP_FAIL.
-     *
-     * \param cfg the SPI bus configuration.
-     * \return ESP_OK or any error code encountered during the inialization.
-     *         Will return ESP_FAIL if called when already installed.
-     */
-    esp_err_t install(const Config& cfg = Config());
-
-    /**
-     * \brief Free the SPI bus. Must be called last, or other methods return ESP_FAIL.
-     *
-     * \return ESP_OK or any error code encountered during the freeing.
-     *         Will return ESP_FAIL if called when not installed.
-     */
-    esp_err_t uninstall();
-
-    uint8_t getChannelsMask() const { return m_channels_mask; } //!< Get the channel mask, specified in Config::channels_mask
+    virtual ~LineSensor();
 
     /**
      * \brief Start the sensor line calibration procedure.
@@ -113,28 +61,6 @@ public:
     void setCalibration(const CalibrationData& data) { m_calibration = data; }
 
     /**
-     * \brief Read values from the chip. Returns values in range <0; LineSensor::MAX_VAL>.
-     *
-     * \param results the results will be APPENDED to this vector.
-     *        It will be unchanged unless the ESP_OK result is returned (except possibly its capacity).
-     *        Between 0 and LineSensor::CHANNELS values are appended, depending on Config::channels_mask.
-     * \param useCalibration scale sensor values with calibration data. In default, uncalibrated state,
-     *        read returns the same values regardless of this param.
-     *        Use startCalibration() and LineSensorCalibrator to generate the calibration data.
-     * \param differential return differential readings, as specified in the MCP3008 datasheet.
-     * \return ESP_OK or any error code encountered during reading.
-     *         Will return ESP_FAIL if called when not installed.
-     */
-    esp_err_t read(std::vector<uint16_t>& results, bool useCalibration = true, bool differential = false) const;
-
-    /**
-     * \brief See the other read(std::vector<uint16_t>&, bool, bool) const method.
-     *
-     * \param dest array MUST be big enough to accomodate all the channels specified by Config::channels_mask!
-     */
-    esp_err_t read(uint16_t *dest, bool useCalibration = true, bool differential = false) const;
-
-    /**
      * \brief Try to determine a black line's position under the sensors.
      *
      * \param white_line the line is white on black background instead of default black on white background.
@@ -156,13 +82,7 @@ public:
 private:
     LineSensor(const LineSensor&) = delete;
 
-    int requestToChannel(int request) const;
-
     CalibrationData m_calibration;
-    spi_device_handle_t m_spi;
-    spi_host_device_t m_spi_dev;
-    bool m_installed;
-    uint8_t m_channels_mask;
 };
 
 /**
@@ -203,7 +123,7 @@ private:
 
     LineSensor& m_sensor;
     LineSensor::CalibrationData m_data;
-    uint16_t m_max[LineSensor::CHANNELS];
+    uint16_t m_max[Driver::CHANNELS];
 };
 
 }; // namespace mcp3008
